feat(automated-checks): wire up assessment card selection store (#6501)
#### Details

This PR integrates the AssessmentCardSelectionStore added in
https://github.com/microsoft/accessibility-insights-web/pull/6457 into
the data converters and new test view container added in
https://github.com/microsoft/accessibility-insights-web/pull/6482,
https://github.com/microsoft/accessibility-insights-web/pull/6480, and
https://github.com/microsoft/accessibility-insights-web/pull/6476.

Wiring up this infrastructure also exposed a bug in which scanning would
be triggered over and over in the new assessment automated checks view
because the `rescanVisualizationWithoutTelemetry` message creator was
being triggered over and over in the `IssuesTable`. This PR also fixes
that bug by refactoring so that we send different messages depending on
the type of data provided (that is, when we're using assessment data, we
trigger an assessment rescan instead of a unified rescan).

##### Motivation

Feature work

##### Context

After this PR there are never any results displayed in the assessment
automated checks test view. This is because we still need to hook up
more action listeners for the `AssessmentCardSelectionStore`, so right
now the store is never populated with any data.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS