feat(automated-checks): remove storeDataKey and make keys consistent across assessment and visualization configuration objects (#6525)
#### Details

This removes the storeDataKey properties from the assessment objects.
This is because we want the visualization configuration objects to be
consistent with the assessment objects.

##### Motivation

Feature work.

##### Context

There's a lot of context to this, so here's a summary:

First off, reasoning.

The assessment object has two keys before this PR: the `storeDataKey`
and the `key`. When an assessment provider looks up a key, it looks for
the `key` property. This is used to store data in the
`AssessmentStoreData`, which is persisted and utilized for backwards
compatibility. The `storeDataKey` is not directly used from the
assessment object but instead is transferred to the visualization
configuration object as the `key` property and later utilized to store
data within the `VisualizationStore` which is only persisted for MV3
reasons.

There is no tangible benefit that can be seen that was gained from
utilizing two different keys here. However, we do run into issues with
this current way of operation. The visualization configuration factory
will return an object that's `key` will align with the
`assessmentObject.storeDataKey` while the assessment providers will
instead return an object which will return an object with `key` aligning
to `assessmentObject.key`. This can cause confusion as we treat both
visualization configuration objects and assessment providers as the
source of truth/a way to identify tests. While we can continue to use
two different keys (by adding a new `someOtherKey` property to
VisualizationConfiguration objects), in this PR, I'm removing them
because there doesn't seem to be a clear benefit to doing so while it
increases confusion within the codebase.

So, we are now using the assessment object `key` property as the
**true** key. This makes sense because that is the key utilized in
assessment store data and is persisted/backwards compatible.

The visualization store uses the `storeDataKey` but is only persisted
per tab. I did make sure to test the scenario where I started an
assessment without these changes and then updated the extension with
these changes: the extension worked as expected and did not cause an
error (because the store state is permissive: i.e. the new state for
that tab had both types of keys and used the new key for changes/state).

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS