feat(automated-checks): populate scan node results data with assessment data (#6526)
#### Details

Prior to these changes, assessment data and was missing some of the
metadata that unified data has about rules and rule failures. This meant
that the the details view cards for the new assessment automated checks
would be missing a lot of the information that the fastpass automated
checks cards include (for example, how to fix text and rule
descriptions). This PR adds the wiring to pass the relevant data through
to the new assessment automated checks view.

##### Motivation

Feature work

##### Context

This PR also includes two small bug fixes discovered while debugging the
new assessment automated checks. These can be separated into separate
PRs if needed, but they're so small I figured the might as well be
lumped together:

- In `main-window-initializer.ts` the new assessment and quick assess
card selection stores were not being properly created and initialized.
This PR adds that logic.
- In `selector-map-helper.ts` there are cases where the
`assessmentCardSelectionStoreData` data is undefined, causing errors.
This PR adds a check to ensure we don't reference it when it's
undefined.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS