build: fix unified unsigned build issues with yarn v3 migration (#6530)
#### Details

This PR attempts to fix two issues with our unified unsigned build that
started occurring with the yarn v3 update:

* Mac and Windows agents were hitting out-of-heap errors during the
test:unified step. This PR adds `--max-old-space-size=512` to that
command to attempt to mitigate this.
* Linux agents were hitting an error during `build:unified:all` where
Yarn would complain about being forced to update a lockfile as part of
an `electron-builder` step. Implemented a workaround for this and a big
comment explaining it in `Gruntfile.js`.
* I verified this locally by setting `$env:CI = 'true'` and running
`yarn build:unified:canary`, verifying that it reproed the issue from
the CI environment, applying the fix in the PR, and then verifying that
the issue no longer occurred

##### Motivation

Fix unified unsigned build

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS