feat(automated-checks): add rule info to scan node results from default rule map (#6539)
#### Details

Uses the default rules to amend rule info to the automated checks.
Removes changes to the test step result that was doing this for us
because that does not cover the backwards compatibility scenario and
would result in a lot of duplicate data.

##### Motivation

Feature work.

##### Context

What if we load an assessment that has a rule that is not included in
the default rules? In this case, we simply do not show that info (since
we do not have it) in the UI.

A further improvement to this approach would be to start storing rule
information onScanCompleted in the AssessmentStore. This could then be
used instead to load the data in combination with the default rule
information.

Screenshots below of what automated checks looks like when loading an
older assessment.
Before:

![image](https://user-images.githubusercontent.com/32555133/227634189-bfd406b2-6c68-4c77-942c-a7630a776343.png)

After:

![image](https://user-images.githubusercontent.com/32555133/227634121-80f624c6-87cc-403c-89e6-bda4d38773e6.png)


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS