feat(automated-checks): fix assessment automated checks visualization refresh behavior (#6540)
#### Details

Fix bug where, upon refresh of the target page, the visualizations in
Assessment > Automated Checks are not re-enabled. This PR updates that
the assessment issues test view to behave similarly to the
RequirementView, which is responsible for handling this behavior (via
the AssessmentViewUpdateHandler).

##### Motivation

Feature work

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS

---------

Co-authored-by: Wahaj Abid <waabid@microsoft.com>