fix: prevent spurious "no tab with id errors" in console (#6570)
#### Details

This PR removes some spurious error spew from the service worker console
along the lines of:

```
Unchecked runtime.lastError: no tab with id: 12345678
sendTabUrlUpdatedAction: tab with ID 12345678 not found, skipping action message
```

The `sendTabUrlUpdatedAction` part happens because
`TargetPageController` is printing a message when it sees a browser
update/navigate event for a tab that it can't successfully query the
browser for information about. However, that case happens commonly in
normal operation - for example, the browser sends such a navigation
event every time you open a new tab because it sends us a navigation
event as it preloads the new tab content in a tab that isn't visible to
our extension. The message looks like an error but it isn't; less
confusing to omit it.

The `Unchecked runtime.lastError` part happens because
`WebextensionBrowserAdapter.getTabAsync` was still using the non-async
`chrome.tabs.getTab` API and manually converting it to a Promise-based
API, but was doing so incorrectly in that it wasn't checking
`chrome.runtime.getLastError` on error like it's supposed to. This PR
fixes that by just converting it to use the async version of the API
directly.

While I was there, I also normalized all the rest of the async
`WebextensionBrowserAdapter` implementations to consistently use `await`
instead of directly returning browser promises. This has no functional
impact but makes it easier to debug these methods - it enables better
stack traces when an error occurs there and also improves the behavior
of breakpoints there.

##### Motivation

Avoid misleading looking errors in console

##### Context

The changes that are particularly interesting to review are in
`target-page-controller.ts` - most of the individual lines changes are
just:
* typings changes from `chrome.tabs.*` types to equivalent `Tab.*` types
to accomadate the use of the async API.
* updating the name of the API from `getTabAsync` to just `getTab` (for
consistency; none of the other async browser adapter methods use the
`*Async` naming convention)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS