feature: Stop building unified client (#6665)
#### Details

This is a shallow cut at deprecating building and testing of the unified
client. This is coming in before the actual cleanup feature so that we
can sever the downloads of
`accessibility-insights-for-android-service-bin`, and to clear the path
for dependabot PR's that work fine in web but have problems in unified.

Included in this PR:
- Exclude unified from yaml files (shallow pass, later pass will go
deeper)
- Exclude unified from package.json
- Exclude unified from gruntfile (shallow pass)
- Exclude `accessibility-insights-for-android-service-bin` from
dependencies

##### Motivation

Prepare for full removal of Unified client. I've created
https://github.com/microsoft/accessibility-insights-web/releases/tag/before-unified-removal
as a tag for anyone who later wants to resume this work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS