feat(remove-android): Remove Android telemetry types (#6753)
#### Details

Deprecating Android support will touch hundreds of files, so we're
breaking it up into more manageable units. This removes the telemetry
types that were defined for Android. There was an intermediate commit
that resulted from noise in generating a snapshot for a unit test. At
first I thought it was a platform problem, but it may have been a
caching issue. I'll open an issue if I can get a consistent repro, but
it seems to be unrelated to this change.

##### Motivation

Android deprecation feature

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue (part of Android deprecation
feature)
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS