feat(remove-android): Odds and ends (#6760)
#### Details

This cleans up some odds and ends from removing Android:
- Remove C# file types from various exclusion lists (I intentionally
left the `.vs` folder, since it was in the `.gitignore` file before it
was added to `.prettierignore`
- Remove some scraps of the old unified building scripts (calls were
already removed, but not the guts)
- Remove configs to debug unified app from inside vscode
- Remove dependabot policy for electron

##### Motivation

Android deprecation feature (this is potentially the last PR for this
step)

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
There's a weird reference to `playwright` from inside main code. It
feels like a bug, but since it's not part of the deprecation, I'll put
it into a separate PR

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: Android deprecation feature
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS