fix: Update os-specific instructions (#6840)
#### Details

As called out in #5251, we were providing instructions that were
targeting only Windows 10, not Windows 11 or macOS. We were also
referring to "Microsoft Edge" as either "new Microsoft Edge" or
"Microsoft Edge Insider", since that was the appropriate name when the
code was written. Times have changed, so we've updated the following:
- In Assessment 17.1 (title: "High contrast mode"), we have replaced the
Windows 10 instructions with a link that opens a new page containing
instructions for selecting a high contrast theme, with instructions for
Windows 10 or Windows 11. We have also added a note that macOS high
contrast settings are not picked up in the browser (at least it didn't
when I tested it).
- In Assessment 17.2 (title: Resize text), we have replaced the Windows
10 instructions with a link that opens a new page instructions for
setting the logical screen size on both Windows or macOS. The macOS
instructions are less precise because of how macOS settings work,
especially if using an external monitor. For example, on my external 4K
display, the smallest available width is 1920.
- Consistently use "Microsoft Edge" as the displayed browser name

##### Screenshots

High contrast:
Description | Screenshot
--- | ---
Old |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/afc3cb16-8f22-4883-8d97-d88507c53171)
New |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/03ad5151-8332-4f0b-b210-161b0d8ff3b5)
Instructions |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/7cd64995-96df-4d47-ad9a-6a49d0fb1c84)

Text Resize:
Description | Screenshot
--- | ---
Old |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/a06b7a48-bd35-4d4d-ad06-148b87bd12fa)
New |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/a294c6c1-ad2b-42f5-88b7-429a81a41d67)
Instructions |
![image](https://github.com/microsoft/accessibility-insights-web/assets/45672944/5c0ccc22-b9a7-4702-9955-7115674c1b14)

##### Motivation

Address #5251

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
Something to note: The Windows-recommended way to increase text size (in
both Win 10 and Win11) is via the Resize Text option in accessibility.
However, that option does not constrain the viewport, so running the
test that way on a full desktop display doesn't really cover the
scenario that is supposed to be tested here. For that reason, we persist
with the way you'd do this on Win7 and earlier.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: #5651
- [x] Ran `yarn fastpass`
- [ ] Added/updated relevant unit test(s) (and ran `yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS