refactor: remove obsolete persistStoreData flag from PersistentStore, take 2 (#6864)
When we originally introduced `PersistentStore` during the MV2->MV3
transition, it included a `persistStoreData` constructor flag which
allowed us to gradually transition different stores to using persistence
depending on whether we were working with an MV2 or MV3 build. This
functionality is now obsolete - all `PersistentStore`-based stores
always use this with a constant `true` input. This commit removes the
obsolete flag.

This is a takeover of abandoned #6300, which was closed due to the
flag's usage in Unified. Now that Unified is gone, this change is
unblocked.

Noticed while working on #6233

Co-authored-by: Dan Bjorge <danielbj@microsoft.com>