chore: Exclude playwright docker image dependabot updates from ClearlyDefined check (#6917)
#### Details

We use playwright focal docker images for end-to-end tests. This image
is updated by dependabot. The corresponding playwright version that the
image uses will be checked by `ClearlyDefined` when it is updated in
`package.json` but the docker image itself is not included in
`ClearlyDefined` so we can exclude these updates from that check.

##### Motivation

fix the build for #6904 and future similar updates

##### Context


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>