chore: Exclude @types namespace from ClearlyDefined check (#7046)
#### Details

We frequently get dependabot PR's for the `@types` namespace, which are
always devDependencies. To streamline maintenance, this PR automatically
excludes these packages from the ClearlyDefined check. It also updates
the validation script and removes the packages in the `@types` namespace
that we've collected so far.

It also includes 2 analyzer-suggested changes of moving `$null` from the
RHS to the LHS of the comparison to prevent accidentally setting the
value instead of comparing it.

##### Motivation

Streamline dependabot handling

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage
report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS