feat(wcag-2-2): Add links for new resources (#7056)
#### Details

This PR adds the new links called out in the WCAG 2.2 spec. It also adds
the "New for WCAG 2.2" UI that gets automatically added to any test that
links to a `guidelineMetadata` value that includes
`guidanceTags.WCAG_2_2`. The proof this last statement is not in this
PR, but I've validated it locally by adding a new assessment page with
the appropriate tag.

##### Motivation

WCAG 2.2 feature work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
Adding the new pages will be handled separately

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: ADO feature
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic
tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See
`CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS