feat(experimental-tags): Add Preview and Experimental Tags to Title of Cards (#7630)
This commit adds support for "preview" and "experimental" tags for rules that are
being developed.
Tags were also added to the title of the cards so they are visible at a
glance to users.
Most of these files are test updates. The only files relevant to this
actual feature are:

1. src/common/types/store-data/guidance-links.ts
2. src/scanner/map-axe-tags-to-guidance-links.ts
3. src/common/components/cards/rule-resources.tsx
4. src/common/components/cards/rules-with-instances.ts
5. src/reports/components/report-sections/minimal-rule-header.tsx

These small changes revealed a strange issue with the way the result
cards are aria-labeled. The OutcomeChip component within the
minimal-rule-header was rendering both a visible count number (e.g.,
"3") and
screen reader text (e.g., "3 Failed"), resulting in labels like 33
failed...

Even though the entire OutcomeChip was wrapped with aria-hidden="true"
in the MinimalRuleHeader, the screen reader text was still contributing
to the button's visible text calculation and causing the axe-core HTML
report to actually fail the rule "label-content-name-mismatch" and
thereby fail some e2e tests.

A number of other formatting changes and snapshot updates were made.